Added guards to prevent deleting a null value, giving me a heart attack when re-doing the dungeoninfo seeding. Reverted a version bump change in compilation, idk why it suddenly won't accept the change anymore. But alas, better this way anyways.

+++ app/Models/Brushline.php
@@ -63,7 +63,9 @@ class Brushline extends Model
         // Delete Path properly if it gets deleted
         static::deleting(function ($item) {
             /** @var $item Brushline */
-            $item->polyline->delete();
+            if ($item->polyline !== null) {
+                $item->polyline->delete();
+            }
         });
     }
 }
