+++ app/Models/Path.php
@@ -42,7 +42,9 @@ class Path extends Model
         // Delete Path properly if it gets deleted
         static::deleting(function ($item) {
             /** @var $item Path */
-            $item->polyline->delete();
+            if ($item->polyline !== null) {
+                $item->polyline->delete();
+            }
         });
     }
 }
